lint&build fixes
